style: util에 있던 상태관리 로직을 sliceUtil 파일로 이동
컴포넌트에서 사용하는 util과 redux에서 사용하는 util을 구분하는 게 좋을
것 같다고 판단하여 변경
